apps/trigger/src/python/requirements.txt
yt-dlp>=2024.10.0
apify-client>=1.7.0
anthropic>=0.40.0
python-dotenv>=1.0.0
